+++ src/server/__tests__/utl_jest.spec.js
@@ -1,0 +1,29 @@
+import {
+  cleanPath,
+} from '../utl.js'
+
+describe('cleanPath', () => {
+
+  test('root directory /', () => {
+    expect(cleanPath('/')).toBe('/')
+    expect(cleanPath('//')).toBe('/')
+    expect(cleanPath('///')).toBe('/')
+  })
+
+  test('other directories', () => {
+    expect(cleanPath('/aa/dir//')).toBe('/aa/dir/')
+    expect(cleanPath('aa/dir/')).toBe('/aa/dir/')
+    expect(cleanPath('/aa//b/c/dir//')).toBe('/aa/b/c/dir/')
+    expect(cleanPath('aa/b/c///dir/')).toBe('/aa/b/c/dir/')
+    expect(cleanPath('aa/b/c/dir/')).toBe('/aa/b/c/dir/')
+  })
+
+  test('files', () => {
+    expect(cleanPath('//aa//b/c////foo')).toBe('/aa/b/c/foo')
+    expect(cleanPath('aa//b/c//foo')).toBe('/aa/b/c/foo')
+    expect(cleanPath('aa/b///c/foo')).toBe('/aa/b/c/foo')
+    expect(cleanPath('aa/b/c/foo')).toBe('/aa/b/c/foo')
+    expect(cleanPath('aa/foo')).toBe('/aa/foo')
+    expect(cleanPath('//aa/foo')).toBe('/aa/foo')
+  })
+})